Entre voisins final

+++ Entrevoisins/app/src/main/java/com/openclassrooms/entrevoisins/ui/neighbour_list/NeighbourFragment.java
@@ -1,6 +1,8 @@
 package com.openclassrooms.entrevoisins.ui.neighbour_list;
 
+import android.app.Activity;
 import android.content.Context;
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.DividerItemDecoration;
@@ -12,6 +14,8 @@ import android.view.ViewGroup;
 
 import com.openclassrooms.entrevoisins.R;
 import com.openclassrooms.entrevoisins.di.DI;
+import com.openclassrooms.entrevoisins.events.ActivityDetail;
+import com.openclassrooms.entrevoisins.events.ClickNeighbourEvent;
 import com.openclassrooms.entrevoisins.events.DeleteNeighbourEvent;
 import com.openclassrooms.entrevoisins.model.Neighbour;
 import com.openclassrooms.entrevoisins.service.NeighbourApiService;
@@ -31,6 +35,7 @@ public class NeighbourFragment extends Fragment {
 
     /**
      * Create and return a new instance
+     *
      * @return @{@link NeighbourFragment}
      */
     public static NeighbourFragment newInstance() {
@@ -83,11 +88,26 @@ public class NeighbourFragment extends Fragment {
 
     /**
      * Fired if the user clicks on a delete button
+     *
      * @param event
      */
     @Subscribe
     public void onDeleteNeighbour(DeleteNeighbourEvent event) {
-        mApiService.deleteNeighbour(event.neighbour);
-        initList();
+        if (event.fragPosition == 0) {
+            mApiService.deleteNeighbour(event.neighbour);
+            mApiService.getFavoritesNeighbours().contains(event.neighbour);
+            {
+                mApiService.deleteFavoritesNeighbours(event.neighbour);
+            }
+            initList();
+        }
+    }
+
+    @Subscribe
+    public void onClickNeighbourEvent(ClickNeighbourEvent event) {
+        Intent intent = new Intent(getActivity(), ActivityDetail.class);
+        intent.putExtra("Neighbour", event.neighbour);
+        intent.putExtra("Position", event.position);
+        startActivity(intent);
     }
 }
